Adding Object Page, navigation and functions for object operations
