+++ README.md
@@ -1,15 +1,21 @@
 # Weaviate Cluster Operations 🔍
 
-Interact with and manage Weaviate Cluster operations. This app provides tools to inspect shards, view collections & tenants, explore schemas, and analyze cluster statistics.
+Interact with and manage Weaviate Cluster operations. This app provides tools to inspect shards, view collections & tenants, explore schemas, analyze cluster statistics and interact with objects.
 
 [![Go to the WebApp](https://static.streamlit.io/badges/streamlit_badge_black_white.svg)](https://weaviatecluster.streamlit.app/)
 
 ### Features
 
-	•	Shards: View shard details across nodes.
-	•	Collections & Tenants: Explore collections and their tenant configurations.
-	•	Schema: Fetch and view the schema of your Weaviate cluster.
+	•	Shards & Nodes: View shard details across nodes as well as Node details.
+	•	Collections & Tenants: Aggregate and view collections and their tenants.
+	•   Collections Configuration: Explore a collection configurations.
+	•	Schema: Fetch and view the schema configuration of your Weaviate cluster.
 	•	Statistics: Analyze cluster synchronization and node statistics.
+	•	Metadata: View cluster metadata & modules.
+	•	Consistency: Analyze Shards for incosistency.
+	•	Object Operations:
+		•	Fetch an object data.
+		•	Analyze consistency of an object across nodes (Hard coded 11 nodes for now as max)
 
 ## How to run it on your own machine
 
@@ -46,4 +52,4 @@ Steps to Run
    $ streamlit run streamlit_app.py
    ```
 
-	4.	Open the app in your browser: The app will typically run at http://localhost:8501.
+	4.	Open the app in your browser: The app will typically run at http://localhost:8501 (Also ticking Local Cluster checkbox allow you to use Local Cluter without need to have Cloud one).
